OBPIH-6627 Move OutboundImportStep to const
