+++ src/js/hooks/outboundImport/useOutboundImportForm.js
@@ -10,9 +10,9 @@ import { useSelector } from 'react-redux';
 import fulfillmentApi from 'api/services/FulfillmentApi';
 import packingListApi from 'api/services/PackingListApi';
 import notification from 'components/Layout/notifications/notification';
-import OutboundImportStep from 'components/stock-movement-wizard/outboundImport/OutboundImportStep';
 import { STOCK_MOVEMENT_URL } from 'consts/applicationUrls';
 import NotificationType from 'consts/notificationTypes';
+import OutboundImportStep from 'consts/OutboundImportStep';
 import { DateFormat } from 'consts/timeFormat';
 import useOutboundImportValidation from 'hooks/outboundImport/useOutboundImportValidation';
 import useQueryParams from 'hooks/useQueryParams';
